mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-ollama>=0.3.7",
    "langgraph>=0.6.6",
    "mcp>=1.13.1",
    "utils>=1.0.2",
]
